Move enums out of catalogs. Update tests. This will be reflected in the documentation.
