+++ src/scooze/database/core.py
@@ -1,11 +1,27 @@
+from enum import StrEnum
 from typing import Any
 
 from bson import ObjectId
 from pymongo import ReturnDocument
 from pymongo.results import DeleteResult
-from scooze.catalogs import DbCollection
 from scooze.database.mongo import db
+from scooze.enum import ExtendedEnum
 from scooze.utils import to_lower_camel
+
+# region Database Enums
+
+
+class DbCollection(ExtendedEnum, StrEnum):
+    """
+    Collections in the scooze database.
+    """
+
+    CARDS = "cards"
+    DECKS = "decks"
+
+
+# endregion
+
 
 # region Single document
