+++ tests/routers/conftest.py
@@ -4,7 +4,7 @@ import pytest
 from bson import ObjectId
 from mongomock import Collection, MongoClient
 from scooze.card import OracleCard
-from scooze.catalogs import DbCollection
+from scooze.database.core import DbCollection
 from scooze.deck import DeckPart
 from scooze.models.card import CardModelData
 from scooze.models.deck import DeckModelIn
