+++ tests/test_deck.py
@@ -4,8 +4,8 @@ from copy import deepcopy
 
 import pytest
 from scooze.card import OracleCard
-from scooze.catalogs import DecklistFormatter, Format, InThe
-from scooze.deck import Deck
+from scooze.catalogs import Format
+from scooze.deck import Deck, DecklistFormatter, InThe
 from scooze.deckpart import DeckDiff, DeckPart
 from scooze.utils import DictDiff
